Introduce static class hierarchy.
`Closure` is not a functor, since we cannot map arbitrary functions
over it. But it sure looks like one, and an applicative one at that.
What we can do is map /static pointers to/ arbitrary functions over
it. `Closure` is not just an applicative functor, it's also a monad,
as well as a comonad, if again we limit the function space to those
functions that can be statically pointed -- to. In fact an entire
hierarchy of classes mirroring the standard classes can be defined,
where the only difference lies in the fact that all pure values' types
appearing in negative position must be a proof of /static/-ness.